Use correct data type in get'er.

+++ src/filters/filter_mgr.cxx
@@ -77,8 +77,8 @@ void Filter_init() {
     
     pyPropertyNode remote_link_node = pyGetNode("/config/remote_link", true);
     pyPropertyNode logging_node = pyGetNode("/config/logging", true);
-    remote_link_skip = remote_link_node.getDouble("filter_skip");
-    logging_skip = logging_node.getDouble("filter_skip");
+    remote_link_skip = remote_link_node.getLong("filter_skip");
+    logging_skip = logging_node.getLong("filter_skip");
 
     // traverse configured modules
     pyPropertyNode group_node = pyGetNode("/config/filters", true);
